src/main/java/edu/iis/mto/serverloadbalancer/ServerLoadBalancer.java
package edu.iis.mto.serverloadbalancer;

public class ServerLoadBalancer {

    public void balance(Server[] servers, Vm[] vms) {
        Server lessLoadedServer = null;
        for (Server server : servers) {
            if (lessLoadedServer == null || server.currentLoadPercentage < lessLoadedServer.currentLoadPercentage) {
                lessLoadedServer = server;
            }
        }
        for (Vm vm : vms) {
            lessLoadedServer.addVm(vm);

        }
    }

}
